// types/teacher.ts
// export interface Teacher {
//   id?: string
//   name: string
//   email: string
//   phone: string
//   status: string
//   instituteId: string
// }

// types/teacher.ts
export interface Teacher {
  id?: string;  // ⬅ optional now
  name: string;
  email: string;
  phone: string
  instituteId: string;
  status: string;
}